+++ titiler/core/titiler/core/__init__.py
@@ -1,0 +1,10 @@
+"""titiler.core"""
+
+from . import dependencies, errors, factory, routing, utils, version  # noqa
+from .factory import (  # noqa
+    BaseTilerFactory,
+    MultiBandTilerFactory,
+    MultiBaseTilerFactory,
+    TilerFactory,
+)
+from .version import __version__  # noqa