+++ titiler/templates/__init__.py
@@ -1,7 +1,0 @@
-"""titiler.templates."""
-
-import pathlib
-
-from starlette.templating import Jinja2Templates
-
-templates = Jinja2Templates(directory=str(pathlib.Path(__file__).parent))